Split template extraction out of entity_filtering and cache template scans
